implementations/discord/requirements.txt
discord.py>=2.4.0
kraken-agent>=0.1.0
python-dotenv>=0.21.0